Moved classes into correct packages
