+++ app/src/main/java/com/amaze/filemanager/fragments/preference_fragments/ColorPref.java
@@ -24,8 +24,8 @@ import android.widget.Toast;
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.amaze.filemanager.R;
 import com.amaze.filemanager.activities.PreferencesActivity;
-import com.amaze.filemanager.fragments.preference_fragments.specialpreferences.ColorPickerDialog;
-import com.amaze.filemanager.fragments.preference_fragments.specialpreferences.InvalidablePreferenceCategory;
+import com.amaze.filemanager.ui.dialogs.ColorPickerDialog;
+import com.amaze.filemanager.ui.views.preference.InvalidablePreferenceCategory;
 import com.amaze.filemanager.utils.Utils;
 import com.amaze.filemanager.utils.color.ColorPreference;
 import com.amaze.filemanager.utils.color.ColorUsage;
